Move Rect into its own module. Move matrix module into its own file. Move Range ops impls to the range module.
