+++ src/position/mod.rs
@@ -3,10 +3,14 @@ use {CharacterCache, Theme, Ui};
 use widget;
 
 pub use self::range::Range;
+pub use self::rect::Rect;
+pub use self::rect::is_over as is_over_rect;
 pub use self::matrix::Matrix;
 
 
+pub mod matrix;
 pub mod range;
+pub mod rect;
 
 
 /// An alias over the Scalar type used throughout Conrod.
@@ -677,540 +681,3 @@ impl Margin {
     }
 }
 
-
-
-impl ::std::ops::Add<Range> for Range {
-    type Output = Range;
-    fn add(self, rhs: Range) -> Range {
-        Range::new(self.start + rhs.start, self.end + rhs.end)
-    }
-}
-
-impl ::std::ops::Sub<Range> for Range {
-    type Output = Range;
-    fn sub(self, rhs: Range) -> Range {
-        Range::new(self.start - rhs.start, self.end - rhs.end)
-    }
-}
-
-
-/// Start and end bounds on a single axis.
-pub type Bounds = Range;
-
-/// Defines a Rectangle's bounds across the x and y axes.
-/// This is a conrod-specific Rectangle in that it's designed to help with layout.
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub struct Rect {
-    /// The start and end positions of the Rectangle on the x axis.
-    pub x: Bounds,
-    /// The start and end positions of the Rectangle on the y axis.
-    pub y: Bounds,
-}
-
-
-impl Rect {
-
-    /// Construct a Rect from a given `Point` and `Dimensions`.
-    pub fn from_xy_dim(xy: Point, dim: Dimensions) -> Self {
-        Rect {
-            x: Range::from_pos_and_len(xy[0], dim[0]),
-            y: Range::from_pos_and_len(xy[1], dim[1]),
-        }
-    }
-
-    /// Construct a Rect from the coordinates of two points.
-    pub fn from_corners(a: Point, b: Point) -> Self {
-        let (left, right) = if a[0] < b[0] { (a[0], b[0]) } else { (b[0], a[0]) };
-        let (bottom, top) = if a[1] < b[1] { (a[1], b[1]) } else { (b[1], a[1]) };
-        Rect {
-            x: Range { start: left, end: right },
-            y: Range { start: bottom, end: top },
-        }
-    }
-
-    /// The Rect representing the area in which two Rects overlap.
-    pub fn overlap(self, other: Self) -> Option<Self> {
-        self.x.overlap(other.x).and_then(|x| self.y.overlap(other.y).map(|y| Rect { x: x, y: y }))
-    }
-
-    /// The Rect that encompass the two given sets of Rect.
-    pub fn max(self, other: Self) -> Self {
-        Rect {
-            x: self.x.max(other.x),
-            y: self.y.max(other.y),
-        }
-    }
-
-    /// The position in the middle of the x bounds.
-    pub fn x(&self) -> Scalar {
-        self.x.middle()
-    }
-
-    /// The position in the middle of the y bounds.
-    pub fn y(&self) -> Scalar {
-        self.y.middle()
-    }
-
-    /// The xy position in the middle of the bounds.
-    pub fn xy(&self) -> Point {
-        [self.x(), self.y()]
-    }
-
-    /// The centered x and y coordinates as a tuple.
-    pub fn x_y(&self) -> (Scalar, Scalar) {
-        (self.x(), self.y())
-    }
-
-    /// The width of the Rect.
-    pub fn w(&self) -> Scalar {
-        self.x.len()
-    }
-
-    /// The height of the Rect.
-    pub fn h(&self) -> Scalar {
-        self.y.len()
-    }
-
-    /// The total dimensions of the Rect.
-    pub fn dim(&self) -> Dimensions {
-        [self.w(), self.h()]
-    }
-
-    /// The width and height of the Rect as a tuple.
-    pub fn w_h(&self) -> (Scalar, Scalar) {
-        (self.w(), self.h())
-    }
-
-    /// Convert the Rect to a `Point` and `Dimensions`.
-    pub fn xy_dim(&self) -> (Point, Dimensions) {
-        (self.xy(), self.dim())
-    }
-
-    /// The Rect's centered coordinates and dimensions in a tuple.
-    pub fn x_y_w_h(&self) -> (Scalar, Scalar, Scalar, Scalar) {
-        let (xy, dim) = self.xy_dim();
-        (xy[0], xy[1], dim[0], dim[1])
-    }
-
-    /// The length of the longest side of the rectangle.
-    pub fn len(&self) -> Scalar {
-        ::utils::partial_max(self.w(), self.h())
-    }
-
-    /// The Rect's lowest y value.
-    pub fn bottom(&self) -> Scalar {
-        self.y.undirected().start
-    }
-
-    /// The Rect's highest y value.
-    pub fn top(&self) -> Scalar {
-        self.y.undirected().end
-    }
-
-    /// The Rect's lowest x value.
-    pub fn left(&self) -> Scalar {
-        self.x.undirected().start
-    }
-
-    /// The Rect's highest x value.
-    pub fn right(&self) -> Scalar {
-        self.x.undirected().end
-    }
-
-    /// The edges of the **Rect** in a tuple (top, bottom, left, right).
-    pub fn l_r_b_t(&self) -> (Scalar, Scalar, Scalar, Scalar) {
-        (self.left(), self.right(), self.bottom(), self.top())
-    }
-
-    /// The left and top edges of the **Rect** along with the width and height.
-    pub fn l_t_w_h(&self) -> (Scalar, Scalar, Scalar, Scalar) {
-        let (w, h) = self.w_h();
-        (self.left(), self.top(), w, h)
-    }
-
-    /// The left and bottom edges of the **Rect** along with the width and height.
-    pub fn l_b_w_h(&self) -> (Scalar, Scalar, Scalar, Scalar) {
-        let (w, h) = self.w_h();
-        (self.left(), self.bottom(), w, h)
-    }
-
-    /// Shift the Rect along the x axis.
-    pub fn shift_x(self, x: Scalar) -> Self {
-        Rect { x: self.x.shift(x), ..self }
-    }
-
-    /// Shift the Rect along the y axis.
-    pub fn shift_y(self, y: Scalar) -> Self {
-        Rect { y: self.y.shift(y), ..self }
-    }
-
-    /// Shift the Rect by the given Point.
-    pub fn shift(self, xy: Point) -> Self {
-        self.shift_x(xy[0]).shift_y(xy[1])
-    }
-
-    /// Does the given point touch the Rectangle.
-    pub fn is_over(&self, xy: Point) -> bool {
-        self.x.is_over(xy[0]) && self.y.is_over(xy[1])
-    }
-
-    /// The Rect with some padding applied to the left edge.
-    pub fn pad_left(self, pad: Scalar) -> Self {
-        Rect { x: self.x.pad_start(pad), ..self }
-    }
-
-    /// The Rect with some padding applied to the right edge.
-    pub fn pad_right(self, pad: Scalar) -> Self {
-        Rect { x: self.x.pad_end(pad), ..self }
-    }
-
-    /// The rect with some padding applied to the bottom edge.
-    pub fn pad_bottom(self, pad: Scalar) -> Self {
-        Rect { y: self.y.pad_start(pad), ..self }
-    }
-
-    /// The Rect with some padding applied to the top edge.
-    pub fn pad_top(self, pad: Scalar) -> Self {
-        Rect { y: self.y.pad_end(pad), ..self }
-    }
-
-    /// The Rect with some padding amount applied to each edge.
-    pub fn pad(self, pad: Scalar) -> Self {
-        let Rect { x, y } = self;
-        Rect { x: x.pad(pad), y: y.pad(pad) }
-    }
-
-    /// The Rect with some padding applied.
-    pub fn padding(self, padding: Padding) -> Self {
-        Rect {
-            x: self.x.pad_ends(padding.left, padding.right),
-            y: self.y.pad_ends(padding.bottom, padding.top),
-        }
-    }
-
-    /// Stretches the closest edge(s) to the given point if the point lies outside of the Rect area.
-    pub fn stretch_to_point(self, point: Point) -> Self {
-        let Rect { x, y } = self;
-        Rect {
-            x: x.stretch_to_value(point[0]),
-            y: y.stretch_to_value(point[1]),
-        }
-    }
-
-    /// Align `self`'s right edge with the left edge of the `other` **Rect**.
-    pub fn left_of(self, other: Self) -> Self {
-        Rect {
-            x: self.x.align_before(other.x),
-            y: self.y,
-        }
-    }
-
-    /// Align `self`'s left edge with the right dge of the `other` **Rect**.
-    pub fn right_of(self, other: Self) -> Self {
-        Rect {
-            x: self.x.align_after(other.x),
-            y: self.y,
-        }
-    }
-
-    /// Align `self`'s top edge with the bottom edge of the `other` **Rect**.
-    pub fn below(self, other: Self) -> Self {
-        Rect {
-            x: self.x,
-            y: self.y.align_before(other.x),
-        }
-    }
-
-    /// Align `self`'s bottom edge with the top edge of the `other` **Rect**.
-    pub fn above(self, other: Self) -> Self {
-        Rect {
-            x: self.x,
-            y: self.y.align_after(other.x),
-        }
-    }
-
-    /// Align `self`'s left edge with the left edge of the `other` **Rect**.
-    pub fn align_left_of(self, other: Self) -> Self {
-        Rect {
-            x: self.x.align_start_of(other.x),
-            y: self.y,
-        }
-    }
-
-    /// Align the middle of `self` with the middle of the `other` **Rect** along the *x* axis.
-    pub fn align_middle_x_of(self, other: Self) -> Self {
-        Rect {
-            x: self.x.align_middle_of(other.x),
-            y: self.y,
-        }
-    }
-
-    /// Align `self`'s right edge with the right edge of the `other` **Rect**.
-    pub fn align_right_of(self, other: Self) -> Self {
-        Rect {
-            x: self.x.align_end_of(other.x),
-            y: self.y,
-        }
-    }
-
-    /// Align `self`'s bottom edge with the bottom edge of the `other` **Rect**.
-    pub fn align_bottom_of(self, other: Self) -> Self {
-        Rect {
-            x: self.x,
-            y: self.y.align_start_of(other.y),
-        }
-    }
-
-    /// Align the middle of `self` with the middle of the `other` **Rect** along the *y* axis.
-    pub fn align_middle_y_of(self, other: Self) -> Self {
-        Rect {
-            x: self.x,
-            y: self.y.align_middle_of(other.y),
-        }
-    }
-
-    /// Align `self`'s top edge with the top edge of the `other` **Rect**.
-    pub fn align_top_of(self, other: Self) -> Self {
-        Rect {
-            x: self.x,
-            y: self.y.align_end_of(other.y),
-        }
-    }
-
-    /// Place `self` along the top left edges of the `other` **Rect**.
-    pub fn top_left_of(self, other: Self) -> Self {
-        self.align_left_of(other).align_top_of(other)
-    }
-
-    /// Place `self` along the top right edges of the `other` **Rect**.
-    pub fn top_right_of(self, other: Self) -> Self {
-        self.align_right_of(other).align_top_of(other)
-    }
-
-    /// Place `self` along the bottom left edges of the `other` **Rect**.
-    pub fn bottom_left_of(self, other: Self) -> Self {
-        self.align_left_of(other).align_bottom_of(other)
-    }
-
-    /// Place `self` along the bottom right edges of the `other` **Rect**.
-    pub fn bottom_right_of(self, other: Self) -> Self {
-        self.align_right_of(other).align_bottom_of(other)
-    }
-
-    /// Place `self` in the middle of the top edge of the `other` **Rect**.
-    pub fn mid_top_of(self, other: Self) -> Self {
-        self.align_middle_x_of(other).align_top_of(other)
-    }
-
-    /// Place `self` in the middle of the bottom edge of the `other` **Rect**.
-    pub fn mid_bottom_of(self, other: Self) -> Self {
-        self.align_middle_x_of(other).align_bottom_of(other)
-    }
-
-    /// Place `self` in the middle of the left edge of the `other` **Rect**.
-    pub fn mid_left_of(self, other: Self) -> Self {
-        self.align_left_of(other).align_middle_y_of(other)
-    }
-
-    /// Place `self` in the middle of the right edge of the `other` **Rect**.
-    pub fn mid_right_of(self, other: Self) -> Self {
-        self.align_right_of(other).align_middle_y_of(other)
-    }
-
-    /// Place `self` directly in the middle of the `other` **Rect**.
-    pub fn middle_of(self, other: Self) -> Self {
-        self.align_middle_x_of(other).align_middle_y_of(other)
-    }
-
-}
-
-impl ::std::ops::Add<Rect> for Rect {
-    type Output = Rect;
-    fn add(self, rhs: Rect) -> Rect {
-        Rect {
-            x: self.x + rhs.x,
-            y: self.y + rhs.y,
-        }
-    }
-}
-
-impl ::std::ops::Sub<Rect> for Rect {
-    type Output = Rect;
-    fn sub(self, rhs: Rect) -> Rect {
-        Rect {
-            x: self.x - rhs.x,
-            y: self.y - rhs.y,
-        }
-    }
-}
-
-
-/// A function to simplify determining whether or not a point `xy` is over a rectangle.
-/// `rect_xy` is the centered coordinatees of the rectangle.
-pub fn is_over_rect(rect_xy: Point, rect_dim: Dimensions, xy: Point) -> bool {
-    Rect::from_xy_dim(rect_xy, rect_dim).is_over(xy)
-}
-
-
-pub mod matrix {
-    use {CharacterCache, Dimension};
-    use super::{Depth, Dimensions, HorizontalAlign, VerticalAlign, Point, Position, Positionable,
-                Scalar, Sizeable};
-    use theme::Theme;
-    use ui::{self, Ui};
-    use widget;
-
-    pub type WidgetNum = usize;
-    pub type ColNum = usize;
-    pub type RowNum = usize;
-    pub type Width = f64;
-    pub type Height = f64;
-    pub type PosX = f64;
-    pub type PosY = f64;
-
-    /// A type to simplify placement of various widgets in a matrix or grid layout.
-    #[derive(Copy, Clone, Debug)]
-    pub struct Matrix {
-        cols: usize,
-        rows: usize,
-        maybe_position: Option<Position>,
-        maybe_x_dimension: Option<Dimension>,
-        maybe_y_dimension: Option<Dimension>,
-        maybe_h_align: Option<HorizontalAlign>,
-        maybe_v_align: Option<VerticalAlign>,
-        cell_pad_w: Scalar,
-        cell_pad_h: Scalar,
-    }
-
-    impl Matrix {
-
-        /// Start building a new position **Matrix**.
-        pub fn new(cols: usize, rows: usize) -> Matrix {
-            Matrix {
-                cols: cols,
-                rows: rows,
-                maybe_position: None,
-                maybe_x_dimension: None,
-                maybe_y_dimension: None,
-                maybe_h_align: None,
-                maybe_v_align: None,
-                cell_pad_w: 0.0,
-                cell_pad_h: 0.0,
-            }
-        }
-
-        /// Produce the matrix with the given cell padding.
-        pub fn cell_padding(mut self, w: Scalar, h: Scalar) -> Matrix {
-            self.cell_pad_w = w;
-            self.cell_pad_h = h;
-            self
-        }
-
-        /// Call the given function for every element in the Matrix.
-        pub fn each_widget<C, F>(self, ui: &mut Ui<C>, mut f: F) where
-            C: CharacterCache,
-            F: FnMut(&mut Ui<C>, WidgetNum, ColNum, RowNum, Point, Dimensions),
-        {
-            use utils::map_range;
-
-            let pos = self.get_position(&ui.theme);
-            let dim = self.get_dim(ui).unwrap_or([0.0, 0.0]);
-            let (h_align, v_align) = self.get_alignment(&ui.theme);
-
-            // If we can infer some new current parent from the position, set that as the current
-            // parent within the given `Ui`.
-            if let Some(id) = ui::parent_from_position(ui, pos) {
-                ui::set_current_parent_idx(ui, id);
-            }
-
-            let xy = ui.calc_xy(None, pos, dim, h_align, v_align, true);
-            let (half_w, half_h) = (dim[0] / 2.0, dim[1] / 2.0);
-            let widget_w = dim[0] / self.cols as f64;
-            let widget_h = dim[1] / self.rows as f64;
-            let x_min = -half_w + widget_w / 2.0;
-            let x_max = half_w + widget_w / 2.0;
-            let y_min = -half_h - widget_h / 2.0;
-            let y_max = half_h - widget_h / 2.0;
-            let mut widget_num = 0;
-            for col in 0..self.cols {
-                for row in 0..self.rows {
-                    let x = xy[0] + map_range(col as f64, 0.0, self.cols as f64, x_min, x_max);
-                    let y = xy[1] + map_range(row as f64, 0.0, self.rows as f64, y_max, y_min);
-                    let w = widget_w - self.cell_pad_w * 2.0;
-                    let h = widget_h - self.cell_pad_h * 2.0;
-                    f(ui, widget_num, col, row, [x, y], [w, h]);
-                    widget_num += 1;
-                }
-            }
-        }
-
-    }
-
-    impl Positionable for Matrix {
-        #[inline]
-        fn position(mut self, pos: Position) -> Self {
-            self.maybe_position = Some(pos);
-            self
-        }
-        #[inline]
-        fn get_position(&self, theme: &Theme) -> Position {
-            self.maybe_position.unwrap_or(theme.position)
-        }
-        #[inline]
-        fn horizontal_align(mut self, h_align: HorizontalAlign) -> Self {
-            self.maybe_h_align = Some(h_align);
-            self
-        }
-        #[inline]
-        fn vertical_align(mut self, v_align: VerticalAlign) -> Self {
-            self.maybe_v_align = Some(v_align);
-            self
-        }
-        #[inline]
-        fn get_horizontal_align(&self, theme: &Theme) -> HorizontalAlign {
-            self.maybe_h_align.unwrap_or(theme.align.horizontal)
-        }
-        #[inline]
-        fn get_vertical_align(&self, theme: &Theme) -> VerticalAlign {
-            self.maybe_v_align.unwrap_or(theme.align.vertical)
-        }
-        #[inline]
-        fn depth(self, _: Depth) -> Self {
-            unimplemented!();
-        }
-        #[inline]
-        fn get_depth(&self) -> Depth {
-            unimplemented!();
-        }
-    }
-
-    impl Sizeable for Matrix {
-        #[inline]
-        fn x_dimension(mut self, w: Dimension) -> Self {
-            self.maybe_x_dimension = Some(w);
-            self
-        }
-        #[inline]
-        fn y_dimension(mut self, h: Dimension) -> Self {
-            self.maybe_y_dimension = Some(h);
-            self
-        }
-        #[inline]
-        fn get_x_dimension<C: CharacterCache>(&self, ui: &Ui<C>) -> Dimension {
-            const DEFAULT_WIDTH: Dimension = Dimension::Absolute(256.0);
-            self.maybe_x_dimension.or_else(|| {
-                ui.theme.widget_style::<widget::matrix::Style>(widget::matrix::KIND)
-                    .map(|default| default.common.maybe_x_dimension.unwrap_or(DEFAULT_WIDTH))
-            }).unwrap_or(DEFAULT_WIDTH)
-        }
-        #[inline]
-        fn get_y_dimension<C: CharacterCache>(&self, ui: &Ui<C>) -> Dimension {
-            const DEFAULT_HEIGHT: Dimension = Dimension::Absolute(256.0);
-            self.maybe_y_dimension.or_else(|| {
-                ui.theme.widget_style::<widget::matrix::Style>(widget::matrix::KIND)
-                    .map(|default| default.common.maybe_y_dimension.unwrap_or(DEFAULT_HEIGHT))
-            }).unwrap_or(DEFAULT_HEIGHT)
-        }
-    }
-
-}
