--- tests/range.rs
@@ -1,10 +1,0 @@
-
-extern crate conrod;
-
-use conrod::Range;
-
-#[test]
-fn from_pos_and_len() {
-}
-
-
